Collapse various exceptions into a single file
